Added utils lib and config parser template structure
