+++ utils/parser/ConfigParser.h
@@ -2,35 +2,25 @@
 
 #include "yaml-cpp/yaml.h"
 
-#include <unordered_map>
+#include "utils.h"
 
 class ConfigParser
 {
 private:
-	struct TaskInfo {
-		std::string cmd;
-		int num_procs;
-		int umask;
-		std::string working_dir;
-		bool auto_start;
-		bool auto_restart;
-		std::vector<int> exit_codes;
-		std::size_t start_retries;
-		std::size_t start_time;
-		std::vector<int> stop_signals;
-		std::size_t stop_time;
-
-		std::string stdout_file;
-		std::string stderr_file;
-		std::unordered_map<std::string, std::string> env;
-	};
+	YAML::Node m_RootNode;
 
 private:
-	YAML::Node m_RootNode;
+	TaskInfo getTask(YAML::Node& node);
 
 public:
 	ConfigParser() = delete;
-	ConfigParser(const std::string& path_to_config);
 	~ConfigParser() = default;
+
+	ConfigParser(std::string_view path_to_config);
+
+	void loadConfig(std::string_view path_to_config);
+
+	std::unordered_map<std::string, TaskInfo> getTasks();
+
 };
